+++ internal/status/changetracker_test.go
@@ -45,7 +45,7 @@ func TestRecordChange_UpdatesCurrentTimeFields(t *testing.T) {
 	now := time.Now()
 	tracker.RecordChange(now, true)
 	assert.Equal(t, now, tracker.lastUpdated)
-	assert.Equal(t, uint64(1), tracker.updateCount)
+	assert.Equal(t, uint32(1), tracker.updateCount)
 }
 
 func TestRecordChange_HeadAndTailSetCorrectly(t *testing.T) {
